Address issue differently than MIN_XFER_TIME_
I was finding that if the latest_dep_earliest_arr trip was also the best
one, then I could get a problem because it would get chosen with high
probability, and then if it was the only one that was good in the
previous link, then that was a problem because it wasn't allowed.
MIN_XFER_TIME_ addressed this but only because stops are typically
defined with a vehicle arriving and then departing at the same time, so
it would prevent this in labeling, but it was a hack, because if the
arrival != departure then it's still a problem.  This is a more lasting
soln: keep that trip id as defining.